requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiofiles
requests
beautifulsoup4
trafilatura
ag2
chromadb
sentence-transformers
ollama
pynput